Check book slug arg is in book repository during prebake (#355)
* Check book slug arg is in book during prebake

* Make force_from_repo local to read_book_slugs
Fix typo from-book -> from-repo in prebake

* Remove outdated comments from docker-entrypoint

+++ dockerfiles/steps/step-prebake.bash
@@ -2,13 +2,17 @@
 parse_book_dir
 
 cp -R "$IO_FETCHED/." "$IO_FETCH_META"
+repo_info="$(set +x && neb parse-repo "$IO_FETCH_META")"
+pages_root="$(set +x && echo "$repo_info" | jq -r '.container.pages_root')"
+media_root="$(set +x && echo "$repo_info" | jq -r '.container.media_root')"
+
+not_found="$(grep -vFxf <(read_book_slugs --from-repo) <(read_book_slugs) || echo -n)"
+[[ -z "$not_found" ]] || die "Slug(s) not found in repository:\n$not_found\n\nValid options are:\n$(read_book_slugs --from-repo)"
+
 neb pre-assemble "$IO_FETCH_META"
 commit_sha="$(set +x && git -C "$IO_FETCH_META" log --format="%h" -1)"
 rm -rf "$IO_FETCH_META/.git"
 
-repo_info="$(set +x && neb parse-repo "$IO_FETCH_META")"
-pages_root="$(set +x && echo "$repo_info" | jq -r '.container.pages_root')"
-media_root="$(set +x && echo "$repo_info" | jq -r '.container.media_root')"
 
 export HACK_CNX_LOOSENESS=1
 # CNX user books do not always contain media directory
